Conflicts:
	mythtv/bindings/python/MythTV/static.py
	mythtv/libs/libmyth/filesysteminfo.cpp
	mythtv/libs/libmyth/libmyth.pro
	mythtv/libs/libmyth/remoteutil.cpp
	mythtv/libs/libmythbase/decodeencode.h
	mythtv/libs/libmythbase/libmythbase.pro
	mythtv/libs/libmythbase/mythcorecontext.cpp
	mythtv/libs/libmythbase/mythsocket.cpp
	mythtv/libs/libmythbase/mythsocket.h
	mythtv/libs/libmythbase/referencecounter.cpp
	mythtv/libs/libmythbase/referencecounter.h
	mythtv/libs/libmythbase/system-unix.cpp
	mythtv/libs/libmythbase/util.cpp
	mythtv/libs/libmythservicecontracts/datacontracts/settingList.h
	mythtv/libs/libmythservicecontracts/libmythservicecontracts.pro
	mythtv/libs/libmythservicecontracts/service.h
	mythtv/libs/libmythtv/dbcheck.cpp
	mythtv/programs/mythbackend/autoexpire.cpp
	mythtv/programs/mythbackend/httpstatus.cpp
	mythtv/programs/mythbackend/main_helpers.cpp
	mythtv/programs/mythbackend/mainserver.cpp
	mythtv/programs/mythbackend/mainserver.h
	mythtv/programs/mythbackend/mythbackend.pro
	mythtv/programs/mythbackend/scheduler.cpp
	mythtv/programs/mythcommflag/main.cpp
	mythtv/programs/mythfrontend/statusbox.cpp
	mythtv/programs/mythjobqueue/main.cpp
	mythtv/programs/mythjobqueue/mythjobqueue.pro
	mythtv/programs/mythtranscode/main.cpp

Temporary commit.  Need to finish the merge before git will allow more
to be pulled in.

+++ mythtv/programs/mythcommflag/HistogramAnalyzer.cpp
@@ -7,6 +7,7 @@
 // MythTV headers
 #include "mythcorecontext.h"
 #include "mythplayer.h"
+#include "mythlogging.h"
 
 // Commercial Flagging headers
 #include "CommDetector2.h"
@@ -51,7 +52,7 @@ readData(QString filename, float *mean, unsigned char *median, float *stddev,
             goto error;
         }
         if (monochromaticval < 0 || monochromaticval > 1 ||
-                medianval < 0 || medianval > UCHAR_MAX ||
+                medianval < 0 || (uint)medianval > UCHAR_MAX ||
                 widthval < 0 || heightval < 0 || colval < 0 || rowval < 0)
         {
             VERBOSE(VB_COMMFLAG, QString(
@@ -67,7 +68,7 @@ readData(QString filename, float *mean, unsigned char *median, float *stddev,
                         .arg(filename).arg(frameno));
                 goto error;
             }
-            if (counter[ii] < 0 || counter[ii] > UCHAR_MAX)
+            if (counter[ii] < 0 || (uint)(counter[ii]) > UCHAR_MAX)
             {
                 VERBOSE(VB_COMMFLAG, QString(
                             "Data out of range in %1: frame %2")
